+++ policyengine_canada/tests/gov/provinces/on/tax/income/credits/lift/low_income_workers_tax_credit.yaml
@@ -2,36 +2,36 @@
   absolute_error_margin: 0.0001
   period: 2023
   input:
-    low_income_workers_tax_credit_base: 875
+    on_low_income_workers_tax_credit_base: 875
     adjusted_family_net_income: 50_000
     is_married: false
   output:
-    low_income_workers_tax_credit: 0
+    on_low_income_workers_tax_credit: 0
 
 - name: Single under reduction amount
   period: 2023
   input:
-    low_income_workers_tax_credit_base: 875
+    on_low_income_workers_tax_credit_base: 875
     adjusted_family_net_income: 32_500
     is_married: false
   output:
-    low_income_workers_tax_credit: 875
+    on_low_income_workers_tax_credit: 875
 
 - name: Single under full reduction amount
   period: 2023
   input:
-    low_income_workers_tax_credit_base: 875
+    on_low_income_workers_tax_credit_base: 875
     adjusted_family_net_income: 40_000
     is_married: false
   output:
-    low_income_workers_tax_credit: 500
+    on_low_income_workers_tax_credit: 500
 
 - name: Married over full reduction amount
   absolute_error_margin: 0.0001
   period: 2023
   input:
-    low_income_workers_tax_credit_base: 875
+    on_low_income_workers_tax_credit_base: 875
     adjusted_family_net_income: 82_500
     is_married: true
   output:
-    low_income_workers_tax_credit: 0
+    on_low_income_workers_tax_credit: 0
