+++ policyengine_canada/tests/gov/provinces/on/tax/income/credits/lift/low_income_workers_tax_credit_base.yaml
@@ -1,53 +1,40 @@
-- name: Wrong province
-  period: 2023
-  input:
-    employment_income: 20_000
-    province: MANITOBA
-  output:
-    low_income_workers_tax_credit_base: 0
-
 - name: No employment income
   period: 2023
   input:
     employment_income: 0
-    province: ONTARIO
   output:
-    low_income_workers_tax_credit_base: 0
+    on_low_income_workers_tax_credit_base: 0
 
 - name: One eligible person with income under full phase-in amount
   absolute_error_margin: 0.0001
   period: 2023
   input:
     employment_income: 10_000
-    province: ONTARIO
-    low_income_workers_tax_credit_eligible_people: 1
+    on_low_income_workers_tax_credit_eligible_people: 1
   output:
-    low_income_workers_tax_credit_base: 505
+    on_low_income_workers_tax_credit_base: 505
 
 - name: One eligible person with income over full phase-in amount
   period: 2023
   input:
     employment_income: 32_500
-    province: ONTARIO
-    low_income_workers_tax_credit_eligible_people: 1
+    on_low_income_workers_tax_credit_eligible_people: 1
   output:
-    low_income_workers_tax_credit_base: 875
+    on_low_income_workers_tax_credit_base: 875
 
 - name: Two eligible people with income over full phase-in amount
   period: 2023
   input:
     employment_income: 65_000
-    province: ONTARIO
-    low_income_workers_tax_credit_eligible_people: 2
+    on_low_income_workers_tax_credit_eligible_people: 2
   output:
-    low_income_workers_tax_credit_base: 1_750
+    on_low_income_workers_tax_credit_base: 1_750
 
 - name: Two eligible people with income under full phase-in amount
   absolute_error_margin: 0.0001
   period: 2023
   input:
     employment_income: 20_000
-    province: ONTARIO
-    low_income_workers_tax_credit_eligible_people: 2
+    on_low_income_workers_tax_credit_eligible_people: 2
   output:
-    low_income_workers_tax_credit_base: 1_010
+    on_low_income_workers_tax_credit_base: 1_010
