+++ policyengine_canada/tests/gov/provinces/on/tax/income/credits/lift/low_income_workers_tax_credit_eligible.yaml
@@ -2,12 +2,22 @@
   period: 2023
   input:
     employment_income: 0
+    province: ONTARIO
   output:
-    low_income_workers_tax_credit_eligible: false
+    on_low_income_workers_tax_credit_eligible: false
 
 - name: Eligible
   period: 2023
   input:
     employment_income: 1
+    province: ONTARIO
   output:
-    low_income_workers_tax_credit_eligible: true
+    on_low_income_workers_tax_credit_eligible: true
+
+- name: Wrong province
+  period: 2023
+  input:
+    employment_income: 1
+    province: MANITOBA
+  output:
+    on_low_income_workers_tax_credit_eligible: false
