+++ policyengine_canada/variables/gov/provinces/on/tax/income/credits/lift/on_low_income_workers_tax_credit_base.py
@@ -1,7 +1,7 @@
 from policyengine_canada.model_api import *
 
 
-class low_income_workers_tax_credit_base(Variable):
+class on_low_income_workers_tax_credit_base(Variable):
     value_type = float
     entity = Person
     label = "Base amount of Ontario Low-Income Workers Tax Credit"
@@ -10,15 +10,12 @@ class low_income_workers_tax_credit_base(Variable):
 
     def formula(person, period, parameters):
         household = person.household
-        province = household("province", period)
-        in_ontario = province == province.possible_values.ONTARIO
         employment_income = person("employment_income", period)
         p = parameters(period).gov.provinces.on.tax.income.credits.lift
         eligible_people = person(
-            "low_income_workers_tax_credit_eligible_people", period
+            "on_low_income_workers_tax_credit_eligible_people", period
         )
-        amount = min_(
+        return min_(
             eligible_people * p.amount,
             employment_income * p.phase_in_rate,
         )
-        return in_ontario * amount
